requirements.txt
dj-database-url==0.4.1
Django==1.10.4
django-rq==0.9.4
psycopg2==2.6.2
PyGithub==1.29
social-auth-app-django==0.0.1
